Fixing the timestamp of both eeg and heart rate to be the same + Changes from int to float for bpm + Change the buttons' size for usability

+++ Assets/Script/ParticipantData1.cs
@@ -51,5 +51,5 @@ public class HeartRateReading
     // in seconds since experiment start
     public float timestamp;
     // beats per minute
-    public int bpm;
+    public float bpm;
 }
